+++ src/components/Profile/UserProfile/UserProfile.adaptive.css
