+++ src/components/Tournaments/TournamentInfo/TournamentInfo.adaptive.css
