+++ src/pages/Tournaments/Tournaments.adaptive.css
